requirements.txt
anthropic[vertex]>=0.7.0
google-cloud-aiplatform
slack-sdk>=3.21.0